+++ core/input_reorder.h
@@ -14,18 +14,18 @@ namespace oidn {
   class InputReorderNode : public Node
   {
   private:
-    ispc::InputReorder impl;
+    std::shared_ptr<Image> color;
+    std::shared_ptr<Image> albedo;
+    std::shared_ptr<Image> normal;
+    std::shared_ptr<Tensor> dst;
+    std::shared_ptr<TransferFunction> transferFunc;
 
-    Image color;
-    Image albedo;
-    Image normal;
-    Ref<Tensor> dst;
-    Ref<TransferFunction> transferFunc;
+    ispc::InputReorder impl;
 
   public:
     InputReorderNode(const Ref<Device>& device,
-                     const Ref<Tensor>& dst,
-                     const Ref<TransferFunction>& transferFunc,
+                     const std::shared_ptr<Tensor>& dst,
+                     const std::shared_ptr<TransferFunction>& transferFunc,
                      bool hdr,
                      bool snorm)
       : Node(device),
@@ -38,26 +38,21 @@ namespace oidn {
              dst->layout == TensorLayout::Chw16c);
       assert(dst->blockSize() == device->getTensorBlockSize());
 
-      impl.dst = *dst;
       setTile(0, 0, 0, 0, 0, 0);
       impl.transferFunc = transferFunc->getImpl();
       impl.hdr = hdr;
       impl.snorm = snorm;
     }
 
-    void setSrc(const Image& color, const Image& albedo, const Image& normal)
+    void setSrc(const std::shared_ptr<Image>& color, const std::shared_ptr<Image>& albedo, const std::shared_ptr<Image>& normal)
     {
-      assert(dst->dims[0] >= color.numChannels()  +
-                             albedo.numChannels() +
-                             normal.numChannels());
+      assert(dst->dims[0] >= (color  ? color->numChannels()  : 0) +
+                             (albedo ? albedo->numChannels() : 0) +
+                             (normal ? normal->numChannels() : 0));
 
       this->color  = color;
       this->albedo = albedo;
       this->normal = normal;
-
-      impl.color  = color;
-      impl.albedo = albedo;
-      impl.normal = normal;
     }
 
     void setTile(int hSrc, int wSrc, int hDst, int wDst, int H, int W)
@@ -72,6 +67,11 @@ namespace oidn {
 
     void execute() override
     {
+      impl.color  = color  ? *color  : Image();
+      impl.albedo = albedo ? *albedo : Image();
+      impl.normal = normal ? *normal : Image();
+      impl.dst = *dst;
+
       assert(impl.H + impl.hSrcBegin <= getHeight());
       assert(impl.W + impl.wSrcBegin <= getWidth());
       assert(impl.H + impl.hDstBegin <= impl.dst.H);
@@ -83,17 +83,17 @@ namespace oidn {
       });
     }
 
-    Ref<Tensor> getDst() const override { return dst; }
+    std::shared_ptr<Tensor> getDst() const override { return dst; }
 
   private:
     int getWidth() const
     {
-      return color ? color.width : (albedo ? albedo.width : normal.width);
+      return color ? color->width : (albedo ? albedo->width : normal->width);
     }
 
     int getHeight() const
     {
-      return color ? color.height : (albedo ? albedo.height : normal.height);
+      return color ? color->height : (albedo ? albedo->height : normal->height);
     }
   };