+++ core/network.h
@@ -8,49 +8,50 @@
 #include "input_reorder.h"
 #include "output_reorder.h"
 #include "progress.h"
+#include "scratch.h"
 
 #pragma once
 
 namespace oidn {
 
-  class Network : public RefCount
+  class Network
   {
   public:
-    Network(const Ref<Device>& device, const std::map<std::string, Ref<Tensor>>& weightsMap);
+    Network(const Ref<Device>& device, const std::map<std::string, std::shared_ptr<Tensor>>& weightsMap);
 
     void execute(Progress& progress);
     double getWorkAmount() const;
 
     // Scratch memory
-    void setScratchSize(size_t size);
-    Ref<Tensor> newTensor(const TensorDesc& desc, ptrdiff_t offset);
-    Image newImage(const ImageDesc& desc, ptrdiff_t offset);
+    void allocScratch(size_t size);
+    std::shared_ptr<Tensor> newTensor(const TensorDesc& desc, ptrdiff_t offset);
+    std::shared_ptr<Image> newImage(const ImageDesc& desc, ptrdiff_t offset);
 
     TensorDesc getInputReorderDesc(const TensorDims& srcDims, int alignment);
 
-    Ref<InputReorderNode> addInputReorder(const Ref<Tensor>& dst,
-                                          const Ref<TransferFunction>& transferFunc,
-                                          bool hdr,
-                                          bool snorm);
+    std::shared_ptr<InputReorderNode> addInputReorder(const std::shared_ptr<Tensor>& dst,
+                                                      const std::shared_ptr<TransferFunction>& transferFunc,
+                                                      bool hdr,
+                                                      bool snorm);
 
-    Ref<OutputReorderNode> addOutputReorder(const Ref<Tensor>& src,
-                                            const Ref<TransferFunction>& transferFunc,
-                                            bool hdr,
-                                            bool snorm);
+    std::shared_ptr<OutputReorderNode> addOutputReorder(const std::shared_ptr<Tensor>& src,
+                                                        const std::shared_ptr<TransferFunction>& transferFunc,
+                                                        bool hdr,
+                                                        bool snorm);
 
     TensorDesc getConvDesc(const std::string& name, const TensorDesc& srcDesc);
-    Ref<Node> addConv(const std::string& name,
-                      const Ref<Tensor>& src,
-                      const Ref<Tensor>& dst,
-                      bool relu = true);
+    std::shared_ptr<Node> addConv(const std::string& name,
+                                  const std::shared_ptr<Tensor>& src,
+                                  const std::shared_ptr<Tensor>& dst,
+                                  bool relu = true);
 
     TensorDesc getPoolDesc(const TensorDesc& srcDesc);
-    Ref<Node> addPool(const Ref<Tensor>& src,
-                      const Ref<Tensor>& dst);
+    std::shared_ptr<Node> addPool(const std::shared_ptr<Tensor>& src,
+                                  const std::shared_ptr<Tensor>& dst);
 
     TensorDesc getUpsampleDesc(const TensorDesc& srcDesc);
-    Ref<Node> addUpsample(const Ref<Tensor>& src,
-                          const Ref<Tensor>& dst);
+    std::shared_ptr<Node> addUpsample(const std::shared_ptr<Tensor>& src,
+                                      const std::shared_ptr<Tensor>& dst);
 
     TensorDesc getConcatDesc(const std::vector<TensorDesc>& srcDescs);
 
@@ -60,14 +61,12 @@ namespace oidn {
     Ref<Device> device;
     int K; // block size of blocked tensor layouts
 
-    std::vector<Ref<Node>> nodes;
-    std::map<std::string, Ref<Tensor>> weightsMap;
+    std::vector<std::shared_ptr<Node>> nodes;
+    std::map<std::string, std::shared_ptr<Tensor>> weightsMap;
+    Ref<ScratchBuffer> scratch;
 
-    Ref<Buffer> scratch;
-    size_t scratchSize = 0;
-
-    Ref<Tensor> padWeights(const Ref<Tensor>& src);
-    Ref<Tensor> padBias(const Ref<Tensor>& src);
+    std::shared_ptr<Tensor> padWeights(const std::shared_ptr<Tensor>& src);
+    std::shared_ptr<Tensor> padBias(const std::shared_ptr<Tensor>& src);
   };
 
 } // namespace oidn