+++ core/reorder.h
@@ -13,13 +13,13 @@ namespace oidn {
   class ReorderNode : public DNNLNode
   {
   private:
-    Ref<Tensor> src;
-    Ref<Tensor> dst;
+    std::shared_ptr<Tensor> src;
+    std::shared_ptr<Tensor> dst;
 
   public:
     ReorderNode(const Ref<Device>& device,
-                const Ref<Tensor>& src,
-                const Ref<Tensor>& dst)
+                const std::shared_ptr<Tensor>& src,
+                const std::shared_ptr<Tensor>& dst)
       : DNNLNode(device),
         src(src), dst(dst)
     {
@@ -28,7 +28,7 @@ namespace oidn {
               {DNNL_ARG_DST, dst->mem}};
     }
 
-    Ref<Tensor> getDst() const override { return dst; }
+    std::shared_ptr<Tensor> getDst() const override { return dst; }
   };
 
 #endif