+++ core/unet.h
@@ -5,6 +5,7 @@
 
 #include "filter.h"
 #include "network.h"
+#include "scratch.h"
 #include "color.h"
 
 namespace oidn {
@@ -22,11 +23,11 @@ namespace oidn {
     static constexpr int overlap         = round_up(receptiveField / 2, alignment); // required spatial overlap between tiles in pixels
 
     // Images
-    Image color;
-    Image albedo;
-    Image normal;
-    Image output;
-    Image outputTemp; // required for in-place tiled filtering
+    std::shared_ptr<Image> color;
+    std::shared_ptr<Image> albedo;
+    std::shared_ptr<Image> normal;
+    std::shared_ptr<Image> output;
+    std::shared_ptr<Image> outputTemp; // required for in-place tiled filtering
 
     // Options
     bool hdr = false;
@@ -34,7 +35,7 @@ namespace oidn {
     bool directional = false;
     float inputScale = std::numeric_limits<float>::quiet_NaN();
     bool cleanAux = false;
-    int maxMemoryMB = 6000; // approximate maximum memory usage in MBs
+    int maxMemoryMB = 3000; // approximate maximum memory usage in MBs
 
     // Image dimensions
     int H = 0;            // image height
@@ -46,10 +47,10 @@ namespace oidn {
     bool inplace = false; // indicates whether input and output buffers overlap
 
     // Network
-    Ref<Network> net;
-    Ref<InputReorderNode> inputReorder;
-    Ref<OutputReorderNode> outputReorder;
-    Ref<TransferFunction> transferFunc;
+    std::unique_ptr<Network> net;
+    std::shared_ptr<InputReorderNode> inputReorder;
+    std::shared_ptr<OutputReorderNode> outputReorder;
+    std::shared_ptr<TransferFunction> transferFunc;
 
     // Weights
     struct
@@ -69,7 +70,7 @@ namespace oidn {
     Data userWeights;
 
     explicit UNetFilter(const Ref<Device>& device);
-    virtual Ref<TransferFunction> makeTransferFunc() = 0;
+    virtual std::shared_ptr<TransferFunction> getTransferFunc() = 0;
 
   public:
     void setData(const std::string& name, const Data& data) override;
@@ -96,13 +97,13 @@ namespace oidn {
   public:
     explicit RTFilter(const Ref<Device>& device);
 
-    void setImage(const std::string& name, const Image& image) override;
+    void setImage(const std::string& name, const std::shared_ptr<Image>& image) override;
     void removeImage(const std::string& name) override;
     void set1i(const std::string& name, int value) override;
     int get1i(const std::string& name) override;
   
   protected:
-    Ref<TransferFunction> makeTransferFunc() override;
+    std::shared_ptr<TransferFunction> getTransferFunc() override;
   };
 
   // ---------------------------------------------------------------------------
@@ -114,13 +115,13 @@ namespace oidn {
   public:
     explicit RTLightmapFilter(const Ref<Device>& device);
 
-    void setImage(const std::string& name, const Image& image) override;
+    void setImage(const std::string& name, const std::shared_ptr<Image>& image) override;
     void removeImage(const std::string& name) override;
     void set1i(const std::string& name, int value) override;
     int get1i(const std::string& name) override;
 
   protected:
-    Ref<TransferFunction> makeTransferFunc() override;
+    std::shared_ptr<TransferFunction> getTransferFunc() override;
   };
 
 } // namespace oidn